Changes to consume menu control across remaining menus (#7311)
* Consumed menu control in AvatarContextMenu

* Fixed mobile issues

* Changes for friends menu

* Changes for share menu

* Fixed width issue of avatar context menu

* Fixed confirm dialog issues

* Changes for Party Menu

* Changes for RoomMenu

* Moved emote menu css

* Changes for ready player me menu

* Renamed to AvatarUploadMenu

* Changes for avatar select menu

* fix missing import in SettingMenu

* fix error in room.tsx
