+++ packages/client-core/src/common/components/ConfirmDialog/index.module.scss
@@ -12,3 +12,8 @@
     }
   }
 }
+
+.dialogActions {
+  gap: 16px;
+  padding: 8px 24px;
+}
