+++ packages/client-core/src/common/components/Icons/CrownIcon.tsx
@@ -1,0 +1,22 @@
+import React from 'react'
+
+import SvgIcon, { SvgIconProps } from '@mui/material/SvgIcon'
+
+export const CrownIcon = (props: SvgIconProps) => {
+  return (
+    <SvgIcon {...props}>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="1em"
+        height="1em"
+        preserveAspectRatio="xMidYMid meet"
+        viewBox="0 0 24 24"
+      >
+        <path
+          fill="currentColor"
+          d="M5 16L3 5l5.5 5L12 4l3.5 6L21 5l-2 11H5m14 3c0 .6-.4 1-1 1H6c-.6 0-1-.4-1-1v-1h14v1Z"
+        />
+      </svg>
+    </SvgIcon>
+  )
+}
