+++ packages/client-core/src/common/components/InputRadio/index.module.scss
@@ -9,3 +9,7 @@
     }
   }
 }
+
+.block {
+  flex-direction: column;
+}
