+++ packages/client-core/src/common/components/InputRadio/index.tsx
@@ -23,28 +23,39 @@ interface Props {
   name?: string
   options: InputMenuItem[]
   sx?: SxProps<Theme>
+  type?: 'inline' | 'block'
   value?: unknown
   onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void
 }
 
-const InputRadio = ({ className, disabled, error, id, label, name, options, sx, value, onChange }: Props) => {
+const InputRadio = ({ className, disabled, error, id, label, name, options, sx, type, value, onChange }: Props) => {
   return (
     <Box sx={{ display: 'flex', flexDirection: 'column', mb: 2, ...sx }}>
       <FormControl
-        className={`${styles.radioField} ${className ?? ''}`}
+        className={`${styles.radioField} ${type === 'block' ? styles.block : ''} ${className ?? ''}`}
         error={error ? true : false}
         disabled={disabled}
         sx={{ display: 'flex', flexDirection: 'row', alignItems: 'center' }}
       >
         <FormLabel sx={{ mt: 0.5, mr: 5 }}>{label}</FormLabel>
-        <RadioGroup id={id} name={name} value={value} onChange={onChange} row>
+        <RadioGroup
+          id={id}
+          name={name}
+          row={type !== 'block'}
+          sx={{ width: type === 'block' ? '100%' : undefined }}
+          value={value}
+          onChange={onChange}
+        >
           {options.map((el, index) => (
-            <FormControlLabel
-              key={index}
-              value={el.value}
-              control={<Radio onPointerUp={handleSoundEffect} onPointerEnter={handleSoundEffect} />}
-              label={el.label}
-            />
+            <>
+              <FormControlLabel
+                key={index}
+                value={el.value}
+                control={<Radio onPointerUp={handleSoundEffect} onPointerEnter={handleSoundEffect} />}
+                label={el.label}
+              />
+              {el.overflowContent}
+            </>
           ))}
         </RadioGroup>
       </FormControl>
