+++ packages/client-core/src/common/components/Menu/index.module.scss
@@ -16,5 +16,11 @@
 
 .dialogTitle {
   display: flex;
+  align-items: center;
   padding: 16px 12px;
+}
+
+.dialogActions {
+  gap: 16px;
+  padding: 8px 24px;
 }
