+++ packages/client-core/src/common/components/Menu/index.tsx
@@ -20,6 +20,7 @@ interface Props {
   open: boolean
   actions?: React.ReactNode
   children?: React.ReactNode
+  contentMargin?: string | number
   header?: React.ReactNode
   isPopover?: boolean
   maxWidth?: Breakpoint | false
@@ -37,6 +38,7 @@ const Menu = ({
   open,
   actions,
   children,
+  contentMargin,
   header,
   isPopover,
   maxWidth,
@@ -61,7 +63,11 @@ const Menu = ({
         <DialogTitle className={styles.dialogTitle}>
           {showBackButton && <IconButton icon={<ArrowBack />} sx={{ mr: 1 }} onClick={onBack} />}
 
-          {title && <Typography variant="h6">{title}</Typography>}
+          {title && (
+            <Typography variant="h6" sx={{ ml: showBackButton ? undefined : 1.5 }}>
+              {title}
+            </Typography>
+          )}
 
           {header}
 
@@ -69,10 +75,10 @@ const Menu = ({
         </DialogTitle>
       )}
 
-      <DialogContent>{children}</DialogContent>
+      <DialogContent sx={{ margin: contentMargin }}>{children}</DialogContent>
 
       {(showDefaultActions || actions) && (
-        <DialogActions>
+        <DialogActions className={styles.dialogActions}>
           {showDefaultActions && (
             <>
               <Button type="outlined" onClick={onClose}>
