+++ packages/client-core/src/common/components/common.module.scss
@@ -1,5 +1,6 @@
 .inputField {
   background: var(--inputBackground);
+  border-radius: 4px;
 
   /**
   * This below style is added to fix background issues with chrome autofill.
@@ -54,4 +55,25 @@
 
 .bold {
   font-weight: bold;
+}
+
+.chip {
+  color: var(--textColor);
+  border-color: var(--textColor);
+}
+
+.divider {
+  margin: 30px 70px 20px;
+  border: 0.2px solid var(--buttonOutlined);
+  border-bottom: none;
+}
+
+@media (max-width: 899px) {
+  .hideSm {
+    display: none;
+  }
+
+  .flexWrapReverseSm {
+    flex-wrap: wrap-reverse;
+  }
 }
