+++ packages/client-core/src/user/components/UserMenu/menus/AvatarSelectMenu.tsx
@@ -1,454 +1,164 @@
-import React, { useRef, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { PerspectiveCamera, Scene, WebGLRenderer } from 'three'
 
-import {
-  AVATAR_FILE_ALLOWED_EXTENSIONS,
-  MAX_AVATAR_FILE_SIZE,
-  MIN_AVATAR_FILE_SIZE,
-  REGEX_VALID_URL,
-  THUMBNAIL_FILE_ALLOWED_EXTENSIONS,
-  THUMBNAIL_HEIGHT,
-  THUMBNAIL_WIDTH
-} from '@xrengine/common/src/constants/AvatarConstants'
-import { StaticResourceInterface } from '@xrengine/common/src/interfaces/StaticResourceInterface'
-import multiLogger from '@xrengine/common/src/logger'
-import { AssetLoader } from '@xrengine/engine/src/assets/classes/AssetLoader'
-import { AudioEffectPlayer } from '@xrengine/engine/src/audio/systems/MediaSystem'
-import { AvatarRigComponent } from '@xrengine/engine/src/avatar/components/AvatarAnimationComponent'
-import { Entity } from '@xrengine/engine/src/ecs/classes/Entity'
-import { getOptionalComponent } from '@xrengine/engine/src/ecs/functions/ComponentFunctions'
+import Avatar from '@xrengine/client-core/src/common/components/Avatar'
+import AvatarPreview from '@xrengine/client-core/src/common/components/AvatarPreview'
+import Button from '@xrengine/client-core/src/common/components/Button'
+import commonStyles from '@xrengine/client-core/src/common/components/common.module.scss'
+import IconButton from '@xrengine/client-core/src/common/components/IconButton'
+import Menu from '@xrengine/client-core/src/common/components/Menu'
+import { AvatarInterface } from '@xrengine/common/src/interfaces/AvatarInterface'
+import { AvatarEffectComponent } from '@xrengine/engine/src/avatar/components/AvatarEffectComponent'
+import { Engine } from '@xrengine/engine/src/ecs/classes/Engine'
+import { hasComponent } from '@xrengine/engine/src/ecs/functions/ComponentFunctions'
+import { getState, useHookstate } from '@xrengine/hyperflux'
 
-import { AccountCircle, ArrowBack, CloudUpload, Help, SystemUpdateAlt } from '@mui/icons-material'
-import Button from '@mui/material/Button'
-import InputBase from '@mui/material/InputBase'
-import Paper from '@mui/material/Paper'
-import { styled } from '@mui/material/styles'
-import Tab from '@mui/material/Tab'
-import Tabs from '@mui/material/Tabs'
+import { Check, Close, KeyboardArrowDown, KeyboardArrowUp, PersonAdd } from '@mui/icons-material'
+import Box from '@mui/material/Box'
+import Grid from '@mui/material/Grid'
 
-import IconLeftClick from '../../../../common/components/Icons/IconLeftClick'
-import { AvatarService } from '../../../services/AvatarService'
-import { loadAvatarForPreview, resetAnimationLogic, validate } from '../../Panel3D/helperFunctions'
-import { useRender3DPanelSystem } from '../../Panel3D/useRender3DPanelSystem'
+import { useAuthState } from '../../../services/AuthService'
+import { AvatarService, AvatarState } from '../../../services/AvatarService'
 import styles from '../index.module.scss'
 import { Views } from '../util'
 
-const logger = multiLogger.child({ component: 'client-core:AvatarSelectMenu' })
-
 interface Props {
-  avatarData?: StaticResourceInterface
   changeActiveMenu: Function
-  onAvatarUpload?: () => void
 }
 
-let camera: PerspectiveCamera
-let scene: Scene
-let renderer: WebGLRenderer = null!
-let entity: Entity = null!
-
-const Input = styled('input')({
-  display: 'none'
-})
-
-function a11yProps(index: number) {
-  return {
-    id: `simple-tab-${index}`,
-    'aria-controls': `simple-tabpanel-${index}`
-  }
-}
-
-interface TabPanelProps {
-  children?: React.ReactNode
-  index: number
-  value: number
-}
-
-function TabPanel({ children, value, index }: TabPanelProps) {
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
-    >
-      {value === index && children}
-    </div>
-  )
-}
-
-export const AvatarUploadModal = ({ avatarData, changeActiveMenu, onAvatarUpload }: Props) => {
-  const [selectedFile, setSelectedFile] = useState<any>(null)
-  const [selectedThumbnail, setSelectedThumbnail] = useState<any>(null)
-  const [avatarName, setAvatarName] = useState('')
-  const [error, setError] = useState('')
-  const [avatarModel, setAvatarModel] = useState<any>(null)
-  const [activeSourceType, setActiveSourceType] = useState(1)
-  const [avatarUrl, setAvatarUrl] = useState('')
-  const [thumbnailUrl, setThumbnailUrl] = useState('')
-  const [validAvatarUrl, setValidAvatarUrl] = useState(false)
-  const [selectedThumbnailUrl, setSelectedThumbNailUrl] = useState<any>(null)
-  const [selectedAvatarlUrl, setSelectedAvatarUrl] = useState<any>(null)
-  const panelRef = useRef() as React.MutableRefObject<HTMLDivElement>
-
-  const renderPanel = useRender3DPanelSystem(panelRef)
-  const { entity, camera, scene, renderer } = renderPanel.state
-
-  const loadAvatarByURL = async (objectURL) => {
-    try {
-      resetAnimationLogic(entity.value)
-      const obj = await loadAvatarForPreview(entity.value, objectURL)
-      if (!obj) {
-        setAvatarModel(null!)
-        setError('Failed to load')
-        return
-      }
-      scene.value.add(obj)
-      const error = validate(obj, renderer.value, scene.value, camera.value)
-      const avatarRigComponent = getOptionalComponent(entity.value, AvatarRigComponent)
-      if (avatarRigComponent) {
-        avatarRigComponent.rig.Neck.getWorldPosition(camera.value.position)
-        camera.value.position.y += 0.2
-        camera.value.position.z = 0.6
-      }
-      setError(error)
-      if (error === '') {
-        setAvatarModel(obj)
-        obj.name = 'avatar'
-      }
-    } catch (err) {
-      logger.error(err)
-      setError(err)
-    }
-  }
-
-  const handleChangeSourceType = (event: React.SyntheticEvent, newValue: number) => {
-    setActiveSourceType(newValue)
-  }
-
-  const handleThumbnailUrlChange = (event) => {
-    event.preventDefault()
-    setThumbnailUrl(event.target.value)
-    if (REGEX_VALID_URL.test(event.target.value)) {
-      fetch(event.target.value)
-        .then((res) => res.blob())
-        .then((data) => setSelectedThumbNailUrl(data))
-        .catch((err) => {
-          setError(err.message)
-        })
-    }
-  }
-
-  const handleAvatarUrlChange = async (event) => {
-    event.preventDefault()
-    setAvatarUrl(event.target.value)
-    if (/\.(?:gltf|glb|vrm)/.test(event.target.value) && REGEX_VALID_URL.test(event.target.value)) {
-      setValidAvatarUrl(true)
-      loadAvatarByURL(event.target.value)
-      fetch(event.target.value)
-        .then((res) => res.blob())
-        .then((data) => setSelectedAvatarUrl(data))
-        .catch((err) => {
-          setError(err.message)
-        })
-    } else {
-      setValidAvatarUrl(false)
-    }
-  }
-
-  const [fileSelected, setFileSelected] = useState(false)
-
+const AvatarSelectMenu = ({ changeActiveMenu }: Props) => {
   const { t } = useTranslation()
+  const authState = useAuthState()
+  const avatarId = authState.user?.avatarId?.value
+  const avatarState = useHookstate(getState(AvatarState))
+  const avatarList = avatarState.avatarList.value
 
-  const handleAvatarChange = (e) => {
-    if (!e.target.files.length) return
-    if (e.target.files[0].size < MIN_AVATAR_FILE_SIZE || e.target.files[0].size > MAX_AVATAR_FILE_SIZE) {
-      setError(
-        t('user:avatar.fileOversized', {
-          minSize: MIN_AVATAR_FILE_SIZE / 1048576,
-          maxSize: MAX_AVATAR_FILE_SIZE / 1048576
-        })
-      )
-      return
-    }
+  const [page, setPage] = useState(0)
+  const [selectedAvatar, setSelectedAvatar] = useState<AvatarInterface | undefined>()
 
-    const file = e.target.files[0]
-    const reader = new FileReader()
-    reader.onload = (fileData) => {
-      try {
-        const assetType = AssetLoader.getAssetType(file.name)
-        if (assetType) {
-          const objectURL = URL.createObjectURL(file) + '#' + file.name
-          loadAvatarByURL(objectURL)
-        }
-      } catch (error) {
-        logger.error(error)
-        setError(t('user:avatar.selectValidFile'))
-      }
-    }
+  useEffect(() => {
+    AvatarService.fetchAvatarList()
+  }, [])
 
-    try {
-      reader.readAsArrayBuffer(file)
-      setFileSelected(true)
-      setSelectedFile(e.target.files[0])
-    } catch (error) {
-      logger.error(error)
-      setError(t('user:avatar.selectValidFile'))
+  const setAvatar = (avatarId: string, avatarURL: string, thumbnailURL: string) => {
+    if (hasComponent(Engine.instance.currentWorld.localClientEntity, AvatarEffectComponent)) return
+    if (authState.user?.value) {
+      AvatarService.updateUserAvatarId(authState.user.id.value!, avatarId, avatarURL, thumbnailURL)
     }
   }
 
-  const handleAvatarNameChange = (e) => {
+  const loadNextAvatars = (e) => {
     e.preventDefault()
-    setAvatarName(e.target.value)
+
+    setPage(page + 1)
+    AvatarService.fetchAvatarList('increment')
   }
 
-  const uploadAvatar = async () => {
-    if (avatarModel == null) return
-    const thumbnailBlob = activeSourceType ? selectedThumbnail : selectedThumbnailUrl
-    const avatarBlob = activeSourceType ? selectedFile : selectedAvatarlUrl
+  const loadPreviousAvatars = (e) => {
+    e.preventDefault()
 
-    if (thumbnailBlob == null) {
-      await new Promise<void>((resolve) => {
-        const canvas = document.createElement('canvas')
-        ;(canvas.width = THUMBNAIL_WIDTH), (canvas.height = THUMBNAIL_HEIGHT)
-        const newContext = canvas.getContext('2d')
-        newContext?.drawImage(renderer.value.domElement, 0, 0)
-        canvas.toBlob((blob) => {
-          AvatarService.createAvatar(avatarBlob, blob!, avatarName, false)
-          resolve()
-        })
-      })
-    } else {
-      await AvatarService.createAvatar(avatarBlob, thumbnailBlob, avatarName, false)
+    setPage(page - 1)
+    AvatarService.fetchAvatarList('decrement')
+  }
+
+  const confirmAvatar = () => {
+    if (selectedAvatar && avatarId != selectedAvatar?.name) {
+      setAvatar(
+        selectedAvatar?.id || '',
+        selectedAvatar?.modelResource?.url || '',
+        selectedAvatar?.thumbnailResource?.url || ''
+      )
+      changeActiveMenu(Views.Closed)
     }
+    setSelectedAvatar(undefined)
+  }
 
-    onAvatarUpload && onAvatarUpload()
+  const selectAvatar = (avatarResources: AvatarInterface) => {
+    setSelectedAvatar(avatarResources)
+  }
+
+  const openAvatarSelectMenu = () => {
+    changeActiveMenu(Views.AvatarUpload)
+  }
+
+  const openProfileMenu = () => {
     changeActiveMenu(Views.Profile)
   }
 
-  const handleThumbnailChange = (e) => {
-    if (e.target.files[0].size < MIN_AVATAR_FILE_SIZE || e.target.files[0].size > MAX_AVATAR_FILE_SIZE) {
-      setError(
-        t('user:avatar.fileOversized', {
-          minSize: MIN_AVATAR_FILE_SIZE / 1048576,
-          maxSize: MAX_AVATAR_FILE_SIZE / 1048576
-        })
-      )
-      return
-    }
-
-    try {
-      setSelectedThumbnail(e.target.files[0])
-    } catch (error) {
-      logger.error(error)
-      setError(t('user:avatar.selectValidThumbnail'))
-    }
-  }
-
-  const openAvatarMenu = (e) => {
-    e.preventDefault()
-    changeActiveMenu(Views.AvatarSelect)
-  }
-
-  const uploadButtonEnabled = fileSelected && !error && avatarName.length > 3
-
   return (
-    <div className={styles.avatarUploadPanel}>
-      <div className={styles.avatarHeaderBlock}>
-        <button type="button" className={styles.iconBlock} onClick={openAvatarMenu}>
-          <ArrowBack />
-        </button>
-        <h2>{t('user:avatar.title')}</h2>
-      </div>
-
-      <Button
-        onClick={() => changeActiveMenu(Views.ReadyPlayer)}
-        className={styles.useReadyMeBtn}
-        onPointerUp={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-        onPointerEnter={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-      >
-        {t('user:usermenu.profile.useReadyPlayerMe')}
-      </Button>
-
-      <div
-        id="stage"
-        ref={panelRef}
-        className={styles.stage}
-        style={{ width: THUMBNAIL_WIDTH + 'px', height: THUMBNAIL_HEIGHT + 'px' }}
-      >
-        <div className={styles.legendContainer}>
-          <Help />
-          <div className={styles.legend}>
-            <div>
-              <IconLeftClick />
-              <br />- <span>{t('user:avatar.rotate')}</span>
-            </div>
-            <div>
-              <span className={styles.shiftKey}>Shift</span> + <IconLeftClick />
-              <br />- <span>{t('user:avatar.pan')}</span>
-            </div>
-          </div>
-        </div>
-      </div>
-      {selectedThumbnail != null && (
-        <div className={styles.thumbnailContainer}>
-          <img
-            src={URL.createObjectURL(selectedThumbnail)}
-            alt={selectedThumbnail?.name}
-            className={styles.thumbnailPreview}
-          />
-        </div>
-      )}
-      {thumbnailUrl.length > 0 && (
-        <div className={styles.thumbnailContainer}>
-          <img src={thumbnailUrl} crossOrigin="anonymous" alt="Avatar" className={styles.thumbnailPreview} />
-        </div>
-      )}
-      <Paper className={styles.paper2}>
-        <InputBase
-          sx={{ ml: 1, flex: 1, color: '#fff', fontWeight: '700', fontSize: '16px' }}
-          inputProps={{ 'aria-label': 'avatar url' }}
-          classes={{ input: styles.input }}
-          value={avatarData?.key ?? avatarName}
-          disabled={!!avatarData?.key}
-          id="avatarName"
-          size="small"
-          name="avatarname"
-          onChange={handleAvatarNameChange}
-          onPointerUp={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-          onPointerEnter={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-          placeholder="Avatar Name"
-        />
-      </Paper>
-      {!avatarData && (
-        <>
-          <div>
-            <Tabs
-              value={activeSourceType}
-              onChange={handleChangeSourceType}
-              onPointerUp={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-              onPointerEnter={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-              aria-label="basic tabs example"
-              classes={{ root: styles.tabRoot, indicator: styles.selected }}
+    <Menu
+      open
+      showBackButton
+      title={t('user:avatar.titleSelectAvatar')}
+      onBack={openProfileMenu}
+      onClose={() => changeActiveMenu(Views.Closed)}
+    >
+      <Box className={styles.menuContent}>
+        <Grid sx={{ height: '100%', mb: 2 }} spacing={2} container className={commonStyles.flexWrapReverseSm}>
+          <Grid item md={3} sx={{ width: '100%' }}>
+            <Button
+              fullWidth
+              startIcon={<PersonAdd />}
+              title={t('user:avatar.upload')}
+              type="gradient"
+              onClick={openAvatarSelectMenu}
             >
-              <Tab
-                className={activeSourceType == 0 ? styles.selectedTab : styles.unselectedTab}
-                label="Use URL"
-                {...a11yProps(0)}
-                classes={{ root: styles.tabRoot }}
-              />
-              <Tab
-                className={activeSourceType == 1 ? styles.selectedTab : styles.unselectedTab}
-                label="Upload Files"
-                {...a11yProps(1)}
-              />
-            </Tabs>
-          </div>
-          <TabPanel value={activeSourceType} index={0}>
-            <div className={styles.controlContainer}>
-              <div className={styles.selectBtns} style={{ margin: '14px 0' }}>
-                <Paper className={styles.paper} style={{ marginRight: '8px', padding: '4px 0' }}>
-                  <InputBase
-                    sx={{ ml: 1, flex: 1, fontWeight: '700', fontSize: '16px' }}
-                    placeholder="Paste Avatar Url..."
-                    inputProps={{ 'aria-label': 'avatar url' }}
-                    classes={{ input: styles.input }}
-                    value={avatarUrl}
-                    onChange={handleAvatarUrlChange}
-                    onPointerUp={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                    onPointerEnter={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
+              {t('user:avatar.upload')}
+            </Button>
+
+            <Button
+              className={commonStyles.hideSm}
+              disabled={!selectedAvatar}
+              fullWidth
+              startIcon={<Close />}
+              title={t('user:avatar.cancel')}
+              type="gradient"
+              onClick={() => {
+                setSelectedAvatar(undefined)
+              }}
+            >
+              {t('user:avatar.cancel')}
+            </Button>
+
+            <Button
+              disabled={!selectedAvatar}
+              fullWidth
+              startIcon={<Check />}
+              title={t('user:avatar.confirm')}
+              type="gradient"
+              onClick={confirmAvatar}
+            >
+              {t('user:avatar.confirm')}
+            </Button>
+          </Grid>
+
+          <Grid item md={6} className={commonStyles.hideSm}>
+            <AvatarPreview selectedAvatar={selectedAvatar} />
+          </Grid>
+
+          <Grid item md={3}>
+            <IconButton icon={<KeyboardArrowUp />} sx={{ display: 'none' }} onClick={loadPreviousAvatars} />
+
+            <Grid container spacing={1} sx={{ height: '275px', overflow: 'auto' }}>
+              {avatarList.map((avatar) => (
+                <Grid key={avatar.id} md={12} item>
+                  <Avatar
+                    imageSrc={avatar.thumbnailResource?.url || ''}
+                    isSelected={avatar.name === selectedAvatar?.name}
+                    name={avatar.name}
+                    type="square"
+                    onChange={() => selectAvatar(avatar)}
                   />
-                </Paper>
-                <Paper className={styles.paper} style={{ padding: '4px 0' }}>
-                  <InputBase
-                    sx={{ ml: 1, flex: 1, fontWeight: '700', fontSize: '16px' }}
-                    placeholder="Paste Thumbnail Url..."
-                    inputProps={{ 'aria-label': 'thumbnail url' }}
-                    classes={{ input: styles.input }}
-                    value={thumbnailUrl}
-                    onChange={handleThumbnailUrlChange}
-                    onPointerUp={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                    onPointerEnter={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                  />
-                </Paper>
-              </div>
-              <button
-                type="button"
-                className={styles.uploadBtn}
-                onClick={uploadAvatar}
-                onPointerUp={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                onPointerEnter={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                disabled={!validAvatarUrl}
-                style={{ cursor: !validAvatarUrl ? 'not-allowed' : 'pointer' }}
-              >
-                {t('user:avatar.lbl-upload')}
-                <CloudUpload />
-              </button>
-            </div>
-          </TabPanel>
-          <TabPanel value={activeSourceType} index={1}>
-            {error.length > 0 && (
-              <div className={styles.selectLabelContainer}>
-                <div className={styles.avatarSelectError}>{error}</div>
-              </div>
-            )}
-            <div className={styles.controlContainer}>
-              <div className={styles.selectBtns}>
-                <label htmlFor="contained-button-file" style={{ marginRight: '8px' }}>
-                  <Input
-                    accept={AVATAR_FILE_ALLOWED_EXTENSIONS}
-                    id="contained-button-file"
-                    type="file"
-                    onChange={handleAvatarChange}
-                    onPointerUp={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                    onPointerEnter={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                  />
-                  <Button
-                    variant="contained"
-                    component="span"
-                    classes={{ root: styles.rootBtn }}
-                    endIcon={<SystemUpdateAlt />}
-                  >
-                    {t('user:avatar.avatar')}
-                  </Button>
-                </label>
-                <label htmlFor="contained-button-file-t">
-                  <Input
-                    accept={THUMBNAIL_FILE_ALLOWED_EXTENSIONS}
-                    id="contained-button-file-t"
-                    type="file"
-                    onChange={handleThumbnailChange}
-                    onPointerUp={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                    onPointerEnter={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                  />
-                  <Button
-                    variant="contained"
-                    component="span"
-                    classes={{ root: styles.rootBtn }}
-                    endIcon={<AccountCircle />}
-                  >
-                    {t('user:avatar.lbl-thumbnail')}
-                  </Button>
-                </label>
-              </div>
-              <button
-                type="button"
-                className={styles.uploadBtn}
-                onClick={uploadAvatar}
-                onPointerUp={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                onPointerEnter={() => AudioEffectPlayer.instance.play(AudioEffectPlayer.SOUNDS.ui)}
-                style={{ cursor: uploadButtonEnabled ? 'pointer' : 'not-allowed' }}
-                disabled={!uploadButtonEnabled}
-              >
-                {t('user:avatar.lbl-upload')}
-                <CloudUpload />
-              </button>
-            </div>
-          </TabPanel>
-        </>
-      )}
-    </div>
+                </Grid>
+              ))}
+            </Grid>
+
+            <IconButton icon={<KeyboardArrowDown />} sx={{ display: 'none' }} onClick={loadNextAvatars} />
+          </Grid>
+        </Grid>
+      </Box>
+    </Menu>
   )
 }
 
-export default AvatarUploadModal
+export default AvatarSelectMenu
