--- packages/client-core/src/user/components/UserMenu/menus/PartyMenu.module.scss
@@ -1,220 +1,0 @@
-@import "@xrengine/client-core/src/styles/imports.module.scss";
-
-.menuPanel {
-  position: absolute;
-  padding: 0 30px;
-  z-index: 6;
-  border-radius: 20px;
-  background: var(--popupBackground);
-  width: calc(100% - 20px);
-  max-width: 400px;
-  height: min(450px, calc(100% - 20px));
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  box-shadow: 16px 16px 32px 0 #11111159;
-  color: black;
-  max-height: calc(100vh - 150px);
-  overflow: auto;
-  touch-action: auto;
-
-  * {
-    touch-action: auto;
-  }
-
-  .partyPanel {
-    width: 100%;
-    height: calc(100% - 60px);
-    margin: 30px 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    gap: 10px;
-
-    .panelHeader {
-      font-size: 18px;
-      color: var(--textColor);
-      font-weight: bold;
-      text-align: center;
-    }
-
-    .midBlock {
-      height: 100%;
-      overflow-y: auto;
-      display: flex;
-      flex-direction: column;
-      gap: 8px;
-      justify-content: start;
-
-      .partyUserBlock {
-        width: 100%;
-        padding: 8px;
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        justify-content: space-between;
-
-        .userBlock {
-          display: flex;
-          flex-direction: row;
-          align-items: center;
-          gap: 10px;
-          color: var(--textColor);
-          font-weight: bold;
-          font-size: 15px;
-
-          .profile {
-            width: 35px;
-            height: 35px;
-            background: #0006;
-            border-radius: 100%;
-            overflow: hidden;
-
-            img {
-              width: 100%;
-              height: auto;
-              display: block;
-            }
-          }
-        }
-
-        .admin {
-          color: var(--blueHover);
-          display: block;
-          text-align: center;
-          width: 100px;
-        }
-
-        .kick {
-          width: 50px;
-          min-width: unset;
-          height: 35px;
-          background: var(--blue);
-          text-transform: none;
-          margin: 0;
-          font-size: 14px;
-
-          &:hover {
-            opacity: 0.8;
-          }
-        }
-      }
-
-      .noUserBlock {
-        width: 100%;
-        height: 100%;
-        text-align: center;
-        padding: 20px;
-        color: var(--blueHover);
-      }
-    }
-
-    .actionBlock {
-      display: flex;
-      flex-direction: row;
-      gap: 8px;
-      padding: 8px;
-      justify-content: center;
-      align-items: center;
-
-      .controls {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-
-        & > :not(:last-child) {
-          margin-bottom: 10px;
-        }
-
-        .leaveInviteButtons,
-        .confirmDeleteButtons {
-          display: flex;
-          flex-direction: row;
-          justify-content: space-between;
-
-          & > :first-child {
-            margin-right: 10px;
-          }
-        }
-      }
-
-      .create,
-      .invite,
-      .leave,
-      .startDelete,
-      .confirmDelete,
-      .cancelDelete {
-        width: 100%;
-        height: 35px;
-        background: linear-gradient(90deg, var(--buttonGradientStart), var(--buttonGradientEnd));
-        text-transform: none;
-        margin: 0;
-
-        &:hover {
-          opacity: 0.8;
-        }
-      }
-
-      .leave,
-      .confirm {
-        background: var(--buttonGradientEnd);
-      }
-
-      .emailField {
-        width: 100%;
-        border: var(--textColor);
-
-        * {
-          border-color: var(--textColor);
-        }
-
-        input {
-          color: var(--textColor);
-        }
-
-        .send {
-          color: var(--blueHover);
-
-          &:hover,
-          &:focus-visible {
-            opacity: 0.8;
-            cursor: pointer;
-          }
-        }
-
-        .cancelInvite {
-          color: var(--red);
-
-          &:hover,
-          &:focus-visible {
-            opacity: 0.8;
-            cursor: pointer;
-          }
-        }
-      }
-    }
-  }
-
-  @media screen and (max-width: 500px) {
-    .headerBlock {
-      width: 100% !important;
-      margin-top: 30px;
-    }
-  }
-}
-
-.backDrop {
-  position: relative;
-
-  &::before {
-    content: "";
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    left: 0;
-    top: 0;
-    background-color: var(--blueHover);
-    opacity: 0.15;
-    border-radius: 5px;
-  }
-}
